+++ src/lte/test/lte-test-rr-ff-mac-scheduler.h
@@ -40,14 +40,14 @@ using namespace ns3;
 class LenaRrFfMacSchedulerTestCase : public TestCase
 {
 public:
-  LenaRrFfMacSchedulerTestCase (uint16_t nUser, uint16_t dist, double thrRefDl, double thrRefUl, bool errorModelEnabled);
+  LenaRrFfMacSchedulerTestCase (uint16_t nUser, double dist, double thrRefDl, double thrRefUl, bool errorModelEnabled);
   virtual ~LenaRrFfMacSchedulerTestCase ();
 
 private:
   virtual void DoRun (void);
-  static std::string BuildNameString (uint16_t nUser, uint16_t dist);
+  static std::string BuildNameString (uint16_t nUser, double dist);
   uint16_t m_nUser;
-  uint16_t m_dist;
+  double m_dist;
   double m_thrRefDl;
   double m_thrRefUl;
   bool m_errorModelEnabled;
