+++ src/lte/test/lte-test-tdbet-ff-mac-scheduler.h
@@ -42,14 +42,14 @@ using namespace ns3;
 class LenaTdBetFfMacSchedulerTestCase1 : public TestCase
 {
 public:
-  LenaTdBetFfMacSchedulerTestCase1 (uint16_t nUser, uint16_t dist, double thrRefDl, double thrRefUl,bool errorModelEnabled);
+  LenaTdBetFfMacSchedulerTestCase1 (uint16_t nUser, double dist, double thrRefDl, double thrRefUl,bool errorModelEnabled);
   virtual ~LenaTdBetFfMacSchedulerTestCase1 ();
 
 private:
-  static std::string BuildNameString (uint16_t nUser, uint16_t dist);
+  static std::string BuildNameString (uint16_t nUser, double dist);
   virtual void DoRun (void);
   uint16_t m_nUser;
-  uint16_t m_dist;
+  double m_dist;
   double m_thrRefDl;
   double m_thrRefUl;
   bool m_errorModelEnabled;
@@ -59,14 +59,14 @@ private:
 class LenaTdBetFfMacSchedulerTestCase2 : public TestCase
 {
 public:
-  LenaTdBetFfMacSchedulerTestCase2 (std::vector<uint16_t> dist, std::vector<uint32_t> m_achievableRateDl, std::vector<uint32_t> estThrTdBetUl, bool m_errorModelEnabled);
+  LenaTdBetFfMacSchedulerTestCase2 (std::vector<double> dist, std::vector<uint32_t> m_achievableRateDl, std::vector<uint32_t> estThrTdBetUl, bool m_errorModelEnabled);
   virtual ~LenaTdBetFfMacSchedulerTestCase2 ();
 
 private:
-  static std::string BuildNameString (uint16_t nUser, std::vector<uint16_t> dist);
+  static std::string BuildNameString (uint16_t nUser, std::vector<double> dist);
   virtual void DoRun (void);
   uint16_t m_nUser;
-  std::vector<uint16_t> m_dist;
+  std::vector<double> m_dist;
   std::vector<uint32_t> m_achievableRateDl;
   std::vector<uint32_t> m_estThrTdBetUl;
   bool m_errorModelEnabled;
