+++ src/lte/test/lte-test-tdmt-ff-mac-scheduler.h
@@ -42,14 +42,14 @@ using namespace ns3;
 class LenaTdMtFfMacSchedulerTestCase : public TestCase
 {
 public:
-  LenaTdMtFfMacSchedulerTestCase (uint16_t nUser, uint16_t dist, double thrRefDl, double thrRefUl,bool errorModelEnabled);
+  LenaTdMtFfMacSchedulerTestCase (uint16_t nUser, double dist, double thrRefDl, double thrRefUl,bool errorModelEnabled);
   virtual ~LenaTdMtFfMacSchedulerTestCase ();
 
 private:
-  static std::string BuildNameString (uint16_t nUser, uint16_t dist);
+  static std::string BuildNameString (uint16_t nUser, double dist);
   virtual void DoRun (void);
   uint16_t m_nUser;
-  uint16_t m_dist;
+  double m_dist;
   double m_thrRefDl;
   double m_thrRefUl;
   bool m_errorModelEnabled;
